+++ ui/src/components/collections/docs/ListApiDocs.svelte
@@ -4,6 +4,7 @@
     import CodeBlock from "@/components/base/CodeBlock.svelte";
     import FilterSyntax from "@/components/collections/docs/FilterSyntax.svelte";
     import SdkTabs from "@/components/collections/docs/SdkTabs.svelte";
+    import FieldsQueryParam from "@/components/collections/docs/FieldsQueryParam.svelte";
 
     export let collection;
 
@@ -209,20 +210,7 @@
                 Only the relations to which the request user has permissions to <strong>view</strong> will be expanded.
             </td>
         </tr>
-        <tr>
-            <td id="query-page">fields</td>
-            <td>
-                <span class="label">String</span>
-            </td>
-            <td>
-                Comma separated string of the fields to return in the JSON response
-                <em>(by default returns all fields)</em>.
-                <br />
-                Use <code>*</code> to target all keys from the specific depth level.
-                <br />
-                For example: <code>?fields=*,expand.author.name</code>
-            </td>
-        </tr>
+        <FieldsQueryParam />
         <tr>
             <td id="query-page">skipTotal</td>
             <td>